requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.0
scikit-image>=0.19.0
trimesh>=3.9.0
scipy>=1.7.0
Pillow>=9.0.0
ultralytics>=8.0.0
kornia>=0.6.0